requirements.txt
# Python Requirements for AI Chemistry Research Assistant

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Environment and Configuration
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# AI/ML Libraries
openai==1.3.7
google-generativeai==0.3.2

# Chemistry Libraries
rdkit-pypi==2022.9.5

# Document Processing
pypdf==3.17.1
python-docx==1.1.0
openpyxl==3.1.2

# HTTP Requests
requests==2.31.0

# Optional: For enhanced development
# pytest==7.4.3  # For testing
# black==23.11.0  # For code formatting
# flake8==6.1.0   # For linting